pyproject.toml
[project]
name = "warehouse-management"
version = "0.1.0"
description = "Warehouse management"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "flask>=3.1.1",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
]

[tool.uv]
dev-dependencies = [
    "flake8>=7.1.2",
]
